requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# src/python/cli.py: 7
# src/python/downloader.py: 3
Pillow == 4.2.1

# src/python/cli.py: 3
# src/python/downloader.py: 1
opencv_python == 3.3.0.10

# src/python/cli.py: 8
# src/python/downloader.py: 4
tqdm == 4.19.4
